#230 [fix,design]:마이페이지 기능 수정 및 디자인 수정
