+++ src/app/(navigationBar)/mypage/[userId]/page.tsx
@@ -3,7 +3,7 @@ import React from 'react';
 import { Tables } from '@/lib/types/supabase';
 type Profile = Tables<'profiles'>;
 
-import ProfileDetail from '@/components/mypage/ProfileDetail';
+import ProfileDetail from '@/components/mypage/profile/ProfileDetail';
 
 const ProfilePage = ({
   params: { userId }
